move host/unhost buttons to browse happs
