+++ src/components/Button/Button.module.css
@@ -5,8 +5,3 @@
   border: 1px solid lightgray;
   border-radius: 4px;
 }
-
-.button:hover {
-  background: lightgray;
-  cursor: pointer;
-}
