--- src/pages/HostingOverview/HostingOverview.module.css
@@ -1,90 +1,0 @@
-.container {
-  width: 100%;
-}
-
-.header {
-  display: flex;
-  padding-bottom: 6px;
-  border-bottom: 1px solid lightgray;
-}
-
-.title {
-  font-style: normal;
-  font-weight: bold;
-  font-size: 24px;
-  line-height: 28px;
-  letter-spacing: 0.02em;
-  color: #484848;
-}
-
-.menu-button {
-  margin-left: auto;
-}
-
-.happ-list {
-  padding-top: 20px;
-}
-
-.pricing-button {
-  /* empty */
-}
-
-.happ-row {
-  margin-bottom: 50px;
-}
-
-.happ-row-main {
-  display: flex;
-  justify-content: space-between;
-}
-
-.icon {
-  display: block;
-  flex: 0 0 25px;  
-  height: 25px;
-  border-radius: 5px;
-}
-
-.expansion-arrow {
-  min-width: 45px;
-}
-
-.happ-row-expanded {
-  /* empty */
-}
-
-.details {
-  padding-top: 8px;
-  padding-left: 25px;
-}
-
-.title-row {
-  display: flex;
-  margin-bottom: 5px;
-}
-
-.title {
-  font-weight: 500;
-  font-size: 16px;
-  line-height: 19px;
-  color: #484848;  
-}
-
-.is-hosted {
-  margin-left: auto;
-  font-size: 16px;
-  line-height: 19px;
-  color: #787878;  
-}
-
-.homepage {
-  /* empty */
-}
-
-.description {
-  margin-top: 25px;
-  font-weight: normal;
-  font-size: 16px;
-  line-height: 19px;
-  color: #939393;
-}
