--- src/pages/HostingOverview/index.js
@@ -1,4 +1,0 @@
-import Component from './HostingOverview'
-import connector from './HostingOverview.connector'
-
-export default connector(Component)
